Update old `jest` tests
