+++ app/components/Card/index.test.js
@@ -1,12 +1,18 @@
 import React from "react";
-import { render, cleanup } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import Card from "./index";
 
 describe("<Card />", () => {
-  afterEach(cleanup);
-  it("Render component correctly", () => {
-    expect(() => {
-      render(<Card />);
-    }).not.toThrow();
+  it("", () => {
+    const props = {
+      header: "One Title",
+      avatar: "http://www.imageshack.us/2421.jpg",
+      href: "http://www.github.com/username",
+      name: "username",
+      children: <div>Hey</div>,
+    };
+    render(<Card {...props} />);
+
+    expect(screen.getByText("One Title")).toBeDefined();
   });
 });
