+++ app/components/Loading/index.test.js
@@ -1,12 +1,11 @@
 import React from "react";
-import { render, cleanup } from "@testing-library/react";
+import { render, screen } from "@testing-library/react";
 import Loading from "./index";
 
 describe("<Loading />", () => {
-  afterEach(cleanup);
-  it("Render component correctly", () => {
-    expect(() => {
-      render(<Loading />);
-    }).not.toThrow();
+  it("", () => {
+    render(<Loading />);
+
+    expect(screen.getByText("Loading")).toBeDefined();
   });
 });
