+++ app/components/Nav/index.test.js
@@ -1,9 +1,8 @@
 import React from "react";
-import { render, cleanup } from "@testing-library/react";
+import { render } from "@testing-library/react";
 import Nav from "./index";
 
 describe("<Nav />", () => {
-  afterEach(cleanup);
   it("Render component correctly", () => {
     expect(() => {
       render(<Nav />);
